関係性入力を必須化 & ディレクトリ構造を整理
