+++ lib/view_models/member/member_list_view_model.dart
@@ -1,7 +1,7 @@
 import 'dart:async';
 import 'package:flutter/material.dart';
-import '../models/kindness_giver.dart';
-import '../repositories/kindness_giver_repository.dart';
+import '../../models/kindness_giver.dart';
+import '../../repositories/kindness_giver_repository.dart';
 
 class KindnessGiverListViewModel extends ChangeNotifier {
   final KindnessGiverRepository _repository = KindnessGiverRepository();
@@ -14,7 +14,7 @@ class KindnessGiverListViewModel extends ChangeNotifier {
   bool get isLoading => _isLoading;
   String? get error => _error;
 
-  // 優しさをくれる人一覧の読み込み
+  // メンバー一覧の読み込み
   Future<void> loadKindnessGivers() async {
     _isLoading = true;
     _error = null;
